API-TRADER-TICKER/requirements.txt
Django==4.2.1
django-bootstrap-v5==1.0.11
django-cors-headers==4.3.1
djangorestframework==3.15.1
yfinance==0.2.18
ta==0.10.2